requirements.txt
absl-py==2.0.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
dm-haiku==0.0.11
e3nn-jax==0.20.3
einops==0.7.0
etils==1.5.2
flax==0.7.5
fonttools==4.45.1
fsspec==2023.10.0
future==0.18.3
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.6
importlib-resources==6.1.1
jax==0.4.20
jax-md==0.2.8
jaxlib==0.4.20
jmp==0.0.4
joblib==1.3.2
jraph==0.0.6.dev0
kiwisolver==1.4.5
latexcodec==2.0.1
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.1
monty==2023.11.3
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.3
packaging==23.2
palettable==3.3.3
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybtex==0.24.0
Pygments==2.17.2
pymatgen==2023.11.12
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
spglib==2.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorstore==0.1.50
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uncertainties==3.1.7
urllib3==2.1.0
zipp==3.17.0
